In progress conversion to modern OpenSSL APIs
Modern OpenSSL makes these structures opaque. Refactor the SWIG
definitions to make this clear, and paste in/shim other getters and
setters as needed.

{} and fields are removed so SWIG knows these are opaque.

The union pkey is also no longer needed, since we can call the provided
getters and setters for the RSA field on the PKEY proper. Changes in the
Python wrapper were also made.

I also got changed the key files because -p would have included them
anyways. Sue me, the other guy hard-coded it decades ago.

However, the current solution for rsac is broken; these fields should
NOT be set indivually, but instead in groups. Not doing so will cause
nulls to be set otherwise, which is Bad Awful No-Good.