Adapt OpenSSL wrapper to call combined setters
OpenSSL 1.1.x changed the API so that you can't set indiviual fields all
at once, but rather entire groups of fields with a setter that sets
multiple, as part of the transition to opaque structures. My workaround
sucked because it tried to simulate setting the indiviual fields, but
this didn't work because it would set the others to NULL, even if you
set them later. You _really_ have to set them all at once. Luckily,
PyGale seems to really set them in a row, so it was trivial to make
these call the combined setter instead of having an indiviual setter for
each.